+++ pkg/client/queue/_examples/main.go
@@ -24,7 +24,7 @@ func main() {
 		}),
 		client.NewClient(
 			message.NewSettings(&message.SSettings{
-				FMessageSizeBytes: 1 << 12,
+				FMessageSizeBytes: (1 << 12),
 				FWorkSizeBits:     10,
 			}),
 			asymmetric.NewRSAPrivKey(1024),